refactor: 実行スクリプトと判定が別のモジュールになる
check-cursor-rule-mirrors.ts が entrypoint と判定を 1 ファイルに
抱えていたので、テストが import しても実行されないための import.meta
ガードが要り、その分岐はテストから踏めないので v8 ignore で計測から
外していた。判定を cursor-rule-mirrors.ts に移し、entrypoint を
process.exit(main(...)) だけにしてガードを消した。orchestrate.ts と
orchestrate-decisions.ts、wrangler-types-plugin.ts と
wrangler-types.ts と同じ分け方になる。CI が呼ぶパスは変わらず、判定側は
20 arms として ignore なしでゲート下にある。
